add condition to houses delete action to prevent Primary home from getting deleted

+++ app/controllers/house_controller.rb
@@ -72,8 +72,9 @@ class HouseController < ApplicationController
   end
 
   delete '/houses/:id/delete' do
+    binding.pry
     @house = House.find(params[:id])
-    if @house.user_id == current_user.id
+    if @house.user_id == current_user.id && @house.name != "Primary"
 
       @house.items.each do |item|
         item.destroy
